requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
avro-python3==1.10.2
certifi==2022.12.7
charset-normalizer==2.1.1
confluent-kafka==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
djangorestframework==3.14.0
drf-yasg==1.21.4
elastic-apm==6.15.1
fastavro==1.7.3
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
openai==0.27.0
packaging==22.0
PyMySQL==1.0.2
python-logstash==0.4.8
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.13
wrapt==1.15.0
yarl==1.8.2
